Refactor to separate twitch integration logic
