+++ src/views/GoTV/index.ts
@@ -16,6 +16,6 @@
  */
 
 export * from "./GoTV";
-export * from "./StreamCard";
 export * from "./GoTVIndicator";
 export * from "./twitchLanguageCodes";
+export * from "./useTwitchIntegration";
